ss6

+++ Web54/mindx-images/modules/post/post.controller.js
@@ -6,6 +6,8 @@
 
 const PostModel = require('./post')
 const CommentModel = require('../comment/comment')
+const UserModel = require("../auth/user")
+const jwt = require("jsonwebtoken")
 
 const getAllPosts = async(req, res) => {
     try {
@@ -46,9 +48,16 @@ const getPostById = async(req, res) => {
 };
 const createPosts = async(req, res) => {
     try {
-        const newPostData = req.body;
+        const { user } = req
 
-        const newPost = await PostModel.create(newPostData);
+        const newPostData = req.body; //imgUrl, title, description, createBy
+
+        //cần phải gửi thông tin để định danh người dùng 
+        // username, password để đăng nhập => check người dùng có hay không
+        const newPost = await PostModel.create({
+            ...newPostData,
+            createBy: user._id
+        });
 
         res.send({
             success: 1,
@@ -66,12 +75,29 @@ const createPosts = async(req, res) => {
 };
 const updatePost = async(req, res) => {
     try {
+        //chỉ người tạo mới có quyền sửa post
         const { postId } = req.params
+        const { user } = req;
+
+        // const existedPost = await PostModel.findById(postId)
+        // if (!existedPost) {
+        //     throw new Error("Not found post")
+        // }
+        // if (String(existedPost.createBy) !== String(user._id)) {
+        //     throw new Error("Not have permission")
+        // }
+
+
         const updatePostData = req.body;
 
         //option { new: true } để kết quả trả về document đã đc update
-        const updatePost = await PostModel.findByIdAndUpdate(postId, updatePostData, { new: true });
-        // const updatePost = await PostModel.findOneAndUpdate({_id: postId}, updatePostData, { new: true });
+        // const updatePost = await PostModel.findByIdAndUpdate(postId, updatePostData, { new: true });
+        const updatePost = await PostModel.findOneAndUpdate({ _id: postId, createBy: user._id }, updatePostData, { new: true });
+
+        if (!updatePost) {
+            throw new Error("Not found post")
+        }
+
         res.send({
             success: 1,
             data: updatePost
@@ -88,6 +114,23 @@ const updatePost = async(req, res) => {
 };
 const deletePost = async(req, res) => {
     try {
+        const token = req.headers.authorization;
+        if (!token) {
+            throw new Error("Bạn không có quyền xóa bài post")
+        }
+        const identityData = jwt.verify(token, "web54")
+
+        const userId = identityData.userId
+
+        //xác thực
+        const existedUser = await UserModel.findById(userId)
+        console.log(identityData, existedUser);
+        //findDB => tìm ra thông tin user
+        // phân quyền
+        if (!existedUser && existedUser.role !== "admin") {
+            throw new Error("Bạn không có quyền xóa bài post")
+        }
+
         const { postId } = req.params
 
         const deletePost = await PostModel.findByIdAndDelete(postId);
